requirements.txt
cerebras_cloud_sdk
requests
numpy